rename: exchange-repository.js 파일명 오타 수정
